Fix api and status change error handling
The FBErrorUtility needed to be used to access certain data from the
error instead of trying to access properties on the error itself. The
raw JSON error is now passed up to JavaScript. The user error message
and title are attached when FBErrorUtility says they should be
displayed to the user.

Resolves #19.